Bar graph added and graph functionalities shifted to utils
